Reduce the content in game.py
Move the GameDisplay class into a separated file
